requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
python-dotenv
pandas
catboost
pytest

sqlmodel
passlib
python-jose
pydantic-settings
httpx